requirements.txt
future
scikit-learn
torch==2.0.1
torch_geometric
torchaudio
torchvision
torchmetrics==0.11.4
einops
mne==1.6.0
hydra-core==1.3.2
wandb
termcolor
optuna